requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
openai>=1.55.0
pypdf>=4.0.0
pymupdf>=1.24.0
python-multipart>=0.0.9
azure-ai-documentintelligence>=1.0.0
azure-core>=1.29.0
azure-search-documents>=11.4.0
python-dotenv>=1.0.0
chromadb>=0.4.0
tenacity>=8.0.0


